Split sensornet.c into simple-application.c and neighbor-discovery.c
