+++ src/neighbor-discovery.h
@@ -1,0 +1,4 @@
+#include "contiki.h"
+
+// Define name of the process so it can be started by other processes just by including this file.
+PROCESS_NAME(neighbor_discovery_process);
